Move backend implementations to their own directories.
