+++ backend/kafka/kafka.go
@@ -1,4 +1,4 @@
-package main
+package kafka
 
 import (
 	"context"
